requirements.txt
numpy
pandas
matplotlib
seaborn
opencv-python
scikit-learn
tensorflow
tensorflow-addons
keras
torch
torchvision
streamlit
xgboost
joblib
Pillow